// lib/api/generated/models/projectRequest.ts
/**
 * Generated by orval v8.16.0 🍺
 * Do not edit manually.
 * Heymoa API
 * Heymoa 서버 REST API
 * OpenAPI spec version: 1.0.0
 */

export interface ProjectRequest {
  /**
   * 프로젝트 이름
   * @minLength 1
   * @maxLength 100
   */
  name: string;
  /**
   * 프로젝트 설명
   * @maxLength 1000
   * @nullable
   */
  description?: string | null;
}